pdf_word_processor/requirements.txt
PyMuPDF>=1.23.0
Pillow>=10.0.0
gTTS>=2.4.0
streamlit>=1.28.0

